Move `ShortestPathPattern` to IR module
