+++ community/cypher/ir-3.2/src/main/scala/org/neo4j/cypher/internal/ir/v3_2/ShortestPathPattern.scala
@@ -17,12 +17,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans
+package org.neo4j.cypher.internal.ir.v3_2
 
 import org.neo4j.cypher.internal.frontend.v3_2.ast
-import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternLength, PatternRelationship}
 
-// TODO: Remove ast representation
 final case class ShortestPathPattern(name: Option[IdName], rel: PatternRelationship, single: Boolean)
                                     (val expr: ast.ShortestPaths) {
